Disable remote DNS lookup on 'lshw -version'
Users are unlikely to expect that a 'lshw -version' does a DNS lookup,
and this may cause alarm in environments where you want to view any
external DNS lookup as something alarming that you should immediately
investigate the cause of.

Additionally, in a Linux distro context, this doesn't necessarily help
the user, as the updated version may not have been packaged yet.

Signed-off-by: Stewart Smith <trawets@amazon.com>